+++ yorm/__init__.py
@@ -3,7 +3,7 @@
 import sys
 
 __project__ = 'YORM'
-__version__ = '0.2.1'
+__version__ = '0.3'
 
 VERSION = __project__ + '-' + __version__
 
@@ -14,7 +14,9 @@ if not sys.version_info >= PYTHON_VERSION:  # pragma: no cover (manual test)
 
 try:
     from . import standard, extended, container
-    from .utilities import UUID, store, store_instances, map_attr
+    from .utilities import UUID
+    from .utilities import sync, sync_object, sync_instances, attr
+    from .utilities import update, update_object, update_file
     from .base import Mappable, Converter
 except ImportError:  # pragma: no cover (manual test)
     pass