+++ yorm/test/test_all.py
@@ -3,9 +3,12 @@
 
 """Integration tests for the `yorm` package."""
 
+import time
+import logging
+
 import pytest
 
-from yorm import store, store_instances, map_attr, Converter
+from yorm import sync, attr, Converter
 from yorm.container import Dictionary, List
 from yorm.standard import Object, String, Integer, Float, Boolean
 from yorm.extended import Markdown
@@ -47,25 +50,25 @@ class EmptyDictionary(Dictionary):
     """Sample dictionary container."""
 
 
-@map_attr(key=String)
+@attr(key=String)
 class SingleKeyDictionary(Dictionary):
 
     """Sample dictionary container."""
 
 
-@map_attr(all=Integer)
+@attr(all=Integer)
 class IntegerList(List):
 
     """Sample list container."""
 
 
-@map_attr(status=Boolean, label=String)
+@attr(status=Boolean, label=String)
 class StatusDictionary(Dictionary):
 
     """Sample dictionary container."""
 
 
-@map_attr(all=StatusDictionary)
+@attr(all=StatusDictionary)
 class StatusDictionaryList(List):
 
     """Sample list container."""
@@ -105,10 +108,10 @@ class SampleNested:
         return "<nested {}>".format(id(self))
 
 
-@map_attr(object=EmptyDictionary, array=IntegerList, string=String)
-@map_attr(number_int=Integer, number_real=Float)
-@map_attr(true=Boolean, false=Boolean)
-@store_instances("path/to/{self.category}/{self.name}.yml")
+@attr(object=EmptyDictionary, array=IntegerList, string=String)
+@attr(number_int=Integer, number_real=Float)
+@attr(true=Boolean, false=Boolean)
+@sync("path/to/{self.category}/{self.name}.yml")
 class SampleStandardDecorated:
 
     """Sample class using standard attribute types."""
@@ -130,11 +133,11 @@ class SampleStandardDecorated:
         return "<decorated {}>".format(id(self))
 
 
-@map_attr(string=String, number_real=Float)
-@store_instances("sample.yml", auto=False)
+@attr(string=String, number_real=Float)
+@sync("sample.yml", auto=False)
 class SampleDecoratedNoAuto:
 
-    """Sample class with automatic storate turned off."""
+    """Sample class with automatic storage turned off."""
 
     def __init__(self):
         self.string = ""
@@ -144,7 +147,7 @@ class SampleDecoratedNoAuto:
         return "<no auto {}>".format(id(self))
 
 
-@map_attr(string=String, number_real=Float)
+@attr(string=String, number_real=Float)
 class SampleDecoratedNoPath:
 
     """Sample class with a manually mapped path."""
@@ -157,7 +160,7 @@ class SampleDecoratedNoPath:
         return "<no path {}>".format(id(self))
 
 
-@store_instances("sample.yml")
+@sync("sample.yml")
 class SampleEmptyDecorated:
 
     """Sample class using standard attribute types."""
@@ -177,7 +180,7 @@ class SampleExtended:
         return "<extended {}>".format(id(self))
 
 
-@store_instances("path/to/directory/{UUID}.yml", mapping={'level': Level})
+@sync("path/to/directory/{UUID}.yml", attrs={'level': Level})
 class SampleCustomDecorated:
 
     """Sample class using custom attribute types."""
@@ -193,23 +196,40 @@ class SampleCustomDecorated:
 # tests #######################################################################
 
 
+def refresh_file_modification_times(seconds=1.1):
+    """Sleep to allow file modification times to refresh."""
+    logging.info("delaying for %s second%s...", seconds,
+                 "" if seconds == 1 else "s")
+    time.sleep(seconds)
+
+
 def test_imports():
-    """Verify the package namespace is mapped correctly."""
+    """Verify the package namespace is correct."""
     # pylint: disable=W0404,W0612,W0621
+
     import yorm
 
-    from yorm import UUID, store, store_instances, map_attr
-    from yorm import Mappable, Converter
+    # Constants
+    from yorm import UUID  # filename placeholder
 
-    from yorm.standard import String
-    from yorm.extended import Markdown
-    from yorm.container import List
+    # Classes
+    from yorm import Mappable  # base class for mapped objects
+    from yorm import Converter  # base class for converters
+    from yorm.standard import String  # and others
+    from yorm.extended import Markdown  # and others
+    from yorm.container import List  # and others
 
-    assert store
-    assert Converter
-    assert yorm.standard.Boolean
-    assert yorm.extended.Markdown
-    assert yorm.container.Dictionary
+    # Decorators
+    from yorm import sync  # enables mapping on a class's instance objects
+    from yorm import sync_instances  # alias for the class decorator
+    from yorm import attr  # alternate API to identify mapped attributes
+
+    # Functions
+    from yorm import sync  # enables mapping on an instance object
+    from yorm import sync_object  # alias for the mapping function
+    from yorm import update  # fetch (if necessary) and store a mapped object
+    from yorm import update_object  # fetch (optional force) a mapped object
+    from yorm import update_file  # store a mapped object
 
 
 @integration
@@ -260,6 +280,7 @@ class TestStandard:
         """.strip().replace("        ", "") + '\n' == text
 
         # change file values
+        refresh_file_modification_times()
         text = """
         array: [4, 5, 6]
         'false': null
@@ -286,18 +307,18 @@ class TestStandard:
         """Verify standard attribute types dump/load correctly (function)."""
         tmpdir.chdir()
         _sample = SampleStandard()
-        mapping = {'object': SingleKeyDictionary,
-                   'array': IntegerList,
-                   'string': String,
-                   'number_int': Integer,
-                   'number_real': Float,
-                   'true': Boolean,
-                   'false': Boolean}
-        sample = store(_sample, "path/to/directory/sample.yml", mapping)
+        attrs = {'object': SingleKeyDictionary,
+                 'array': IntegerList,
+                 'string': String,
+                 'number_int': Integer,
+                 'number_real': Float,
+                 'true': Boolean,
+                 'false': Boolean}
+        sample = sync(_sample, "path/to/directory/sample.yml", attrs)
         assert "path/to/directory/sample.yml" == sample.yorm_path
 
         # check defaults
-        assert {'key': ''} == sample.object
+        assert {} == sample.object
         assert [] == sample.array
         assert "" == sample.string
         assert 0 == sample.number_int
@@ -337,9 +358,9 @@ class TestStandard:
         """Verify standard attribute types dump/load correctly (with)."""
         tmpdir.chdir()
         _sample = SampleStandard()
-        mapping = {'string': String,
-                   'number_real': Float}
-        sample = store(_sample, "sample.yml", mapping, auto=False)
+        attrs = {'string': String,
+                 'number_real': Float}
+        sample = sync(_sample, "sample.yml", attrs, auto=False)
 
         # change object values
         with sample:
@@ -374,7 +395,7 @@ class TestStandard:
         assert "" == text
 
         # store value
-        sample.yorm_mapper.store(sample)
+        sample.yorm_mapper.store(sample, sample.yorm_attrs)
         sample.yorm_mapper.auto = True
 
         # check for changed file values
@@ -388,7 +409,7 @@ class TestStandard:
     def test_no_path(self, tmpdir):
         """Verify standard attribute types dump/load correctly (path)."""
         tmpdir.chdir()
-        sample = store(SampleDecoratedNoPath(), "sample.yml")
+        sample = sync(SampleDecoratedNoPath(), "sample.yml")
 
         # change object values
         sample.string = "abc"
@@ -412,9 +433,9 @@ class TestContainers:
         """Verify standard attribute types can be nested."""
         tmpdir.chdir()
         _sample = SampleNested()
-        mapping = {'count': Integer,
-                   'results': StatusDictionaryList}
-        sample = store(_sample, "sample.yml", mapping)
+        attrs = {'count': Integer,
+                 'results': StatusDictionaryList}
+        sample = sync(_sample, "sample.yml", attrs)
 
         # check defaults
         assert 0 == sample.count
@@ -447,6 +468,7 @@ class TestContainers:
         """.strip().replace("        ", "") + '\n' == text
 
         # change file values
+        refresh_file_modification_times()
         text = """
         count: 3
         other: 4.2
@@ -472,6 +494,7 @@ class TestContainers:
         sample = SampleEmptyDecorated()
 
         # change file values
+        refresh_file_modification_times()
         text = """
         object: {'key': 'value'}
         array: [1, '2', '3.0']
@@ -479,8 +502,8 @@ class TestContainers:
         with open(sample.yorm_path, 'w') as stream:
             stream.write(text)
 
-            # (a mapped attribute must be read first to trigger retrieving)
-        sample.yorm_mapper.retrieve(sample)
+        # (a mapped attribute must be read first to trigger retrieving)
+        sample.yorm_mapper.fetch(sample, sample.yorm_attrs)
 
         # check object values
         assert {'key': 'value'} == sample.object
@@ -512,13 +535,14 @@ class TestExtended:
         """Verify extended attribute types dump/load correctly."""
         tmpdir.chdir()
         _sample = SampleExtended()
-        mapping = {'text': Markdown}
-        sample = store(_sample, "path/to/directory/sample.yml", mapping)
+        attrs = {'text': Markdown}
+        sample = sync(_sample, "path/to/directory/sample.yml", attrs)
 
         # check defaults
         assert "" == sample.text
 
         # change object values
+        refresh_file_modification_times()
         sample.text = """
         This is the first sentence. This is the second sentence.
         This is the third sentence.
@@ -535,6 +559,7 @@ class TestExtended:
         """.strip().replace("        ", "") + '\n' == text
 
         # change file values
+        refresh_file_modification_times()
         text = """
         text: |
           This is a
@@ -571,6 +596,7 @@ class TestCustom:
         """.strip().replace("        ", "") + '\n' == text
 
         # change file values
+        refresh_file_modification_times()
         text = """
         level: 1
         """.strip().replace("        ", "") + '\n'
@@ -586,14 +612,16 @@ class TestInit:
 
     """Integration tests for initializing mapped classes."""
 
-    def test_retrieve_from_existing(self, tmpdir):
-        """Verify attributes are loaded from an existing file."""
+    def test_fetch_from_existing(self, tmpdir):
+        """Verify attributes are updated from an existing file."""
         tmpdir.chdir()
         sample = SampleStandardDecorated('sample')
         sample2 = SampleStandardDecorated('sample')
         assert sample2.yorm_path == sample.yorm_path
 
-        # change object values
+        refresh_file_modification_times()
+
+        logging.info("changing values in object 1...")
         sample.object = {'key2': 'value'}
         sample.array = [0, 1, 2]
         sample.string = "Hello, world!"
@@ -602,8 +630,8 @@ class TestInit:
         sample.true = True
         sample.false = False
 
-        # check object values
-        assert {'key2': 'value', 'status': False} == sample2.object
+        logging.info("reading changed values in object 2...")
+        assert 'value' == sample2.object.get('key2')
         assert [0, 1, 2] == sample2.array
         assert "Hello, world!" == sample2.string
         assert 42 == sample2.number_int