+++ yorm/test/test_utilities.py
@@ -8,6 +8,7 @@ from unittest.mock import patch, Mock
 
 from yorm import common
 from yorm import utilities
+from yorm.base import Mappable
 from yorm.standard import Converter
 
 
@@ -52,13 +53,22 @@ class MockConverter4(MockConverter):
     """Sample converter class."""
 
 
+class MockMappable(Mappable):
+
+    """Sample mappable class."""
+
+    yorm_attrs = []
+    yorm_mapper = Mock()
+
+
 # tests #######################################################################
 
 
 @patch('yorm.common.write_text', Mock())
-class TestStore:
+@patch('yorm.common.stamp', Mock())
+class TestSyncObject:
 
-    """Unit tests for the `store` function."""
+    """Unit tests for the `sync_object` function."""
 
     class Sample:
 
@@ -66,66 +76,73 @@ class TestStore:
 
     def test_no_attrs(self):
         """Verify mapping can be enabled with no attributes."""
-        sample = utilities.store(self.Sample(), "sample.yml")
+        sample = utilities.sync(self.Sample(), "sample.yml")
         assert "sample.yml" == sample.yorm_path
         assert {} == sample.yorm_attrs
 
     def test_with_attrs(self):
         """Verify mapping can be enabled with with attributes."""
-        mapping = {'var1': MockConverter}
-        sample = utilities.store(self.Sample(), "sample.yml", mapping)
+        attrs = {'var1': MockConverter}
+        sample = utilities.sync(self.Sample(), "sample.yml", attrs)
         assert "sample.yml" == sample.yorm_path
         assert {'var1': MockConverter} == sample.yorm_attrs
 
     def test_multiple(self):
         """Verify mapping cannot be enabled twice."""
-        sample = utilities.store(self.Sample(), "sample.yml")
+        sample = utilities.sync(self.Sample(), "sample.yml")
         with pytest.raises(common.UseageError):
-            utilities.store(sample, "sample.yml")
+            utilities.sync(sample, "sample.yml")
 
-    @patch('os.path.exists', Mock(return_value=True))
+    @patch('os.path.isfile', Mock(return_value=True))
     @patch('yorm.common.read_text', Mock(return_value="abc: 123"))
     def test_init_existing(self):
         """Verify an existing file is read."""
-        sample = utilities.store(self.Sample(), "sample.yml")
+        sample = utilities.sync(self.Sample(), "sample.yml")
         assert 123 == sample.abc
 
     def test_store(self):
         """Verify store is called when setting an attribute."""
-        mapping = {'var1': MockConverter}
-        sample = utilities.store(self.Sample(), "sample.yml", mapping)
+        attrs = {'var1': MockConverter}
+        sample = utilities.sync(self.Sample(), "sample.yml", attrs)
         with patch.object(sample, 'yorm_mapper') as mock_yorm_mapper:
             setattr(sample, 'var1', None)
-        mock_yorm_mapper.retrieve.assert_never_called()
-        mock_yorm_mapper.store.assert_called_once_with(sample)
+        assert not mock_yorm_mapper.fetch.called
+        mock_yorm_mapper.store.assert_called_once_with(sample, attrs)
 
-    def test_retrieve(self):
-        """Verify retrieve is called when getting an attribute."""
-        mapping = {'var1': MockConverter}
-        sample = utilities.store(self.Sample(), "sample.yml", mapping)
+    def test_fetch(self):
+        """Verify fetch is called when getting an attribute."""
+        attrs = {'var1': MockConverter}
+        sample = utilities.sync(self.Sample(), "sample.yml", attrs)
         with patch.object(sample, 'yorm_mapper') as mock_yorm_mapper:
             getattr(sample, 'var1', None)
-        mock_yorm_mapper.retrieve.assert_called_once_with(sample)
-        mock_yorm_mapper.store.assert_never_called()
+        mock_yorm_mapper.fetch.assert_called_once_with(sample, attrs)
+        assert not mock_yorm_mapper.store.called
 
 
 @patch('yorm.common.create_dirname', Mock())
 @patch('yorm.common.write_text', Mock())
-class TestStoreInstances:
+@patch('yorm.common.stamp', Mock())
+class TestSyncInstances:
 
-    """Unit tests for the `store_instances` decorator."""
+    """Unit tests for the `sync_instances` decorator."""
 
-    @utilities.store_instances("sample.yml")
+    @utilities.sync("sample.yml")
     class SampleDecorated:
 
         """Sample decorated class using a single path."""
 
-    @utilities.store_instances("{UUID}.yml")
+        def __repr__(self):
+            return "<decorated {}>".format(id(self))
+
+    @utilities.sync("{UUID}.yml")
     class SampleDecoratedIdentifiers:
 
         """Sample decorated class using UUIDs for paths."""
 
-    @utilities.store_instances("path/to/{n}.yml", {'n': 'name'})
+        def __repr__(self):
+            return "<decorated w/ UUID {}>".format(id(self))
+
+    @utilities.sync("path/to/{n}.yml", {'n': 'name'})
     class SampleDecoratedAttributes:
 
         """Sample decorated class using an attribute value for paths."""
@@ -133,7 +150,10 @@ class TestStoreInstances:
         def __init__(self, name):
             self.name = name
 
-    @utilities.store_instances("path/to/{self.name}.yml")
+        def __repr__(self):
+            return "<decorated w/ specified attributes {}>".format(id(self))
+
+    @utilities.sync("path/to/{self.name}.yml")
     class SampleDecoratedAttributesAutomatic:
 
         """Sample decorated class using an attribute value for paths."""
@@ -141,8 +161,10 @@ class TestStoreInstances:
         def __init__(self, name):
             self.name = name
 
-    @utilities.store_instances("{self.a}/{self.b}/{c}.yml",
-                               {'self.b': 'b', 'c': 'c'})
+        def __repr__(self):
+            return "<decorated w/ automatic attributes {}>".format(id(self))
+
+    @utilities.sync("{self.a}/{self.b}/{c}.yml", {'self.b': 'b', 'c': 'c'})
     class SampleDecoratedAttributesCombination:
 
         """Sample decorated class using an attribute value for paths."""
@@ -152,7 +174,10 @@ class TestStoreInstances:
             self.b = b
             self.c = c
 
-    @utilities.store_instances("sample.yml", mapping={'var1': MockConverter})
+        def __repr__(self):
+            return "<decorated w/ attributes {}>".format(id(self))
+
+    @utilities.sync("sample.yml", attrs={'var1': MockConverter})
     class SampleDecoratedWithAttributes:
 
         """Sample decorated class using a single path."""
@@ -169,7 +194,7 @@ class TestStoreInstances:
         assert "sample.yml" == sample.yorm_path
         assert ['var1'] == list(sample.yorm_attrs.keys())
 
-    @patch('os.path.exists', Mock(return_value=True))
+    @patch('os.path.isfile', Mock(return_value=True))
     @patch('yorm.common.read_text', Mock(return_value="abc: 123"))
     def test_init_existing(self):
         """Verify an existing file is read."""
@@ -209,60 +234,63 @@ class TestStoreInstances:
         sample = self.SampleDecoratedWithAttributes()
         with patch.object(sample, 'yorm_mapper') as mock_yorm_mapper:
             setattr(sample, 'var1', None)
-        mock_yorm_mapper.retrieve.assert_never_called()
-        mock_yorm_mapper.store.assert_called_once_with(sample)
+        assert not mock_yorm_mapper.fetch.called
+        mock_yorm_mapper.store.assert_called_once_with(sample,
+                                                       sample.yorm_attrs)
 
-    def test_retrieve(self):
-        """Verify retrieve is called when getting an attribute."""
+    def test_fetch(self):
+        """Verify fetch is called when getting an attribute."""
         sample = self.SampleDecoratedWithAttributes()
         with patch.object(sample, 'yorm_mapper') as mock_yorm_mapper:
             getattr(sample, 'var1', None)
-        mock_yorm_mapper.retrieve.assert_called_once_with(sample)
-        mock_yorm_mapper.store.assert_never_called()
+        mock_yorm_mapper.fetch.assert_called_once_with(sample,
+                                                       sample.yorm_attrs)
+        assert not mock_yorm_mapper.store.called
 
 
 @patch('yorm.common.write_text', Mock())
-class TestMapAttr:
+@patch('yorm.common.stamp', Mock())
+class TestAttr:
 
-    """Unit tests for the `map_attr` decorator."""
+    """Unit tests for the `attr` decorator."""
 
-    @utilities.map_attr(var1=MockConverter1, var2=MockConverter2)
-    @utilities.store_instances("sample.yml")
+    @utilities.attr(var1=MockConverter1, var2=MockConverter2)
+    @utilities.sync("sample.yml")
     class SampleDecoratedSingle:
 
-        """Sample decorated class using one `map_attr` decorator."""
+        """Sample decorated class using one `attr` decorator."""
 
-    @utilities.map_attr()
-    @utilities.map_attr(var1=MockConverter1)
-    @utilities.map_attr(var2=MockConverter2, var3=MockConverter3)
-    @utilities.store_instances("sample.yml")
+    @utilities.attr()
+    @utilities.attr(var1=MockConverter1)
+    @utilities.attr(var2=MockConverter2, var3=MockConverter3)
+    @utilities.sync("sample.yml")
     class SampleDecoratedMultiple:
 
-        """Sample decorated class using many `map_attr` decorators."""
+        """Sample decorated class using many `attr` decorators."""
 
-    @utilities.map_attr()
-    @utilities.map_attr(var1=MockConverter1)
-    @utilities.map_attr(var2=MockConverter2, var3=MockConverter3)
-    @utilities.store_instances("sample.yml", mapping={'var0': MockConverter0})
+    @utilities.attr()
+    @utilities.attr(var1=MockConverter1)
+    @utilities.attr(var2=MockConverter2, var3=MockConverter3)
+    @utilities.sync("sample.yml", attrs={'var0': MockConverter0})
     class SampleDecoratedCombo:
 
-        """Sample decorated class using `map_attr` and providing a mapping."""
+        """Sample decorated class using `attr` and providing a mapping."""
 
-    @utilities.store_instances("sample.yml", mapping={'var0': MockConverter0})
-    @utilities.map_attr(var1=MockConverter1)
+    @utilities.sync("sample.yml", attrs={'var0': MockConverter0})
+    @utilities.attr(var1=MockConverter1)
     class SampleDecoratedBackwards:
 
-        """Sample decorated class using one `map_attr` decorator."""
+        """Sample decorated class using one `attr` decorator."""
 
     def test_single(self):
-        """Verify `map_attr` can be applied once."""
+        """Verify `attr` can be applied once."""
         sample = self.SampleDecoratedSingle()
         expected = {'var1': MockConverter1,
                     'var2': MockConverter2}
         assert expected == sample.yorm_attrs
 
     def test_multiple(self):
-        """Verify `map_attr` can be applied many times."""
+        """Verify `attr` can be applied many times."""
         sample = self.SampleDecoratedMultiple()
         expected = {'var1': MockConverter1,
                     'var2': MockConverter2,
@@ -270,7 +298,7 @@ class TestMapAttr:
         assert expected == sample.yorm_attrs
 
     def test_combo(self):
-        """Verify `map_attr` can be applied an existing mapping."""
+        """Verify `attr` can be applied an existing mapping."""
         sample = self.SampleDecoratedCombo()
         expected = {'var0': MockConverter0,
                     'var1': MockConverter1,
@@ -279,11 +307,97 @@ class TestMapAttr:
         assert expected == sample.yorm_attrs
 
     def test_backwards(self):
-        """Verify `map_attr` can be applied before `store_instances`."""
+        """Verify `attr` can be applied before `sync`."""
         sample = self.SampleDecoratedBackwards()
         expected = {'var0': MockConverter0,
                     'var1': MockConverter1}
         assert expected == sample.yorm_attrs
+
+
+class TestUpdate:
+
+    """Unit tests for the `update` function."""
+
+    def test_update(self):
+        """Verify the object and file are updated."""
+        instance = MockMappable()
+        instance.yorm_mapper.reset_mock()
+
+        utilities.update(instance)
+
+        assert instance.yorm_mapper.fetch.called
+        assert instance.yorm_mapper.store.called
+
+    def test_update_object_only(self):
+        """Verify only the object is updated."""
+        instance = MockMappable()
+        instance.yorm_mapper.reset_mock()
+
+        utilities.update(instance, store=False)
+
+        assert instance.yorm_mapper.fetch.called
+        assert not instance.yorm_mapper.store.called
+
+    def test_update_file_only(self):
+        """Verify only the file is updated."""
+        instance = MockMappable()
+        instance.yorm_mapper.reset_mock()
+
+        utilities.update(instance, fetch=False)
+
+        assert not instance.yorm_mapper.fetch.called
+        assert instance.yorm_mapper.store.called
+
+    def test_update_wrong_base(self):
+        """Verify an exception is raised with the wrong base."""
+        instance = Mock()
+
+        with pytest.raises(common.UseageError):
+            utilities.update(instance)
+
+
+class TestUpdateObject:
+
+    """Unit tests for the `update_object` function."""
+
+    def test_update(self):
+        """Verify only the object is updated."""
+        instance = MockMappable()
+        instance.yorm_mapper.reset_mock()
+
+        utilities.update_object(instance)
+
+        assert instance.yorm_mapper.fetch.called
+        assert not instance.yorm_mapper.store.called
+
+    def test_update_wrong_base(self):
+        """Verify an exception is raised with the wrong base."""
+        instance = Mock()
+
+        with pytest.raises(common.UseageError):
+            utilities.update_object(instance)
+
+
+class TestUpdateFile:
+
+    """Unit tests for the `update_file` function."""
+
+    def test_update(self):
+        """Verify only the file is updated."""
+        instance = MockMappable()
+        instance.yorm_mapper.reset_mock()
+
+        utilities.update_file(instance)
+
+        assert not instance.yorm_mapper.fetch.called
+        assert instance.yorm_mapper.store.called
+
+    def test_update_wrong_base(self):
+        """Verify an exception is raised with the wrong base."""
+        instance = Mock()
+
+        with pytest.raises(common.UseageError):
+            utilities.update_file(instance)
 
 
 if __name__ == '__main__':